requirements.txt
# 🤖 Telegram API
telethon==1.35.0

# 🎨 Обработка изображений
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4

# 🧠 AI/LLM интеграция
openai==1.12.0
requests==2.31.0

# ⚙️ Утилиты
python-dotenv==1.0.1
aiofiles==23.2.1

# 📊 Мониторинг и логирование
colorlog==6.8.2

